EnderecoServicces sendo ajustada agora para trazer uma matrix com enderecos proximos e fazer listagem

+++ src/main/java/com/coinnect/CoinNect/services/ContratanteServices.java
@@ -1,18 +1,18 @@
 package com.coinnect.CoinNect.services;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Set;
 
 import com.coinnect.CoinNect.model.dtos.ContratanteDTO;
 import com.coinnect.CoinNect.model.dtos.ContratoDTO;
-import com.coinnect.CoinNect.model.enums.StatusContrato;
 
 public interface ContratanteServices {
 
 	ContratanteDTO registrarContrantePorCpf(ContratanteDTO contratante) throws Exception; 
 	ContratanteDTO registrarContrantePorCnpj(ContratanteDTO contratanteCnpj) throws Exception;
 	ContratoDTO oferecerContratoPrestador(ContratoDTO ofertaContrato, Long prestadorId) throws Exception;
-	Double fazerAvaliacaoContratante(Double nota, Long contratanteId) throws Exception;
+	BigDecimal fazerAvaliacaoContratante(Double nota, Long contratanteId) throws Exception;
 	Set<ContratoDTO> verTodosContratos(Long contratanteId) throws Exception;
 	List<ContratanteDTO> listarContrantesBemAvaliados() throws Exception;
 	List<ContratanteDTO> listarContrantesMalAvaliados() throws Exception;
